refactor: move types from lib::accounting::server_consumption to wire
