+++ lib/src/accounting/server_consumption.rs
@@ -2,33 +2,14 @@ use crate::common::{request, SerializableNone};
 use crate::error::ApiError;
 use anyhow::Context;
 use chrono::{DateTime, FixedOffset};
+use lrzcc_wire::accounting::{
+    ServerConsumptionAll, ServerConsumptionFlavors, ServerConsumptionProject,
+    ServerConsumptionServer, ServerConsumptionUser,
+};
 use reqwest::blocking::Client;
 use reqwest::{Method, StatusCode, Url};
-use serde::{Deserialize, Serialize};
-use std::collections::HashMap;
 use std::fmt::Debug;
 use std::rc::Rc;
-
-pub type ServerConsumptionFlavors = HashMap<String, f64>;
-pub type ServerConsumptionServer = ServerConsumptionFlavors;
-
-#[derive(Clone, Debug, Deserialize, Serialize)]
-pub struct ServerConsumptionUser {
-    pub total: ServerConsumptionFlavors,
-    pub servers: HashMap<String, ServerConsumptionServer>,
-}
-
-#[derive(Clone, Debug, Deserialize, Serialize)]
-pub struct ServerConsumptionProject {
-    pub total: ServerConsumptionFlavors,
-    pub users: HashMap<String, ServerConsumptionUser>,
-}
-
-#[derive(Clone, Debug, Deserialize, Serialize)]
-pub struct ServerConsumptionAll {
-    pub total: ServerConsumptionFlavors,
-    pub projects: HashMap<String, ServerConsumptionProject>,
-}
 
 #[derive(Debug)]
 pub struct ServerConsumptionRequest {
