+++ wire/src/accounting/mod.rs
@@ -1,3 +1,5 @@
+mod server_consumption;
 mod server_state;
 
+pub use server_consumption::*;
 pub use server_state::*;
